% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cOligos.R
\name{cOligos}
\alias{cOligos}
\title{Concatenate Oligo (Probe) Sequences}
\usage{
cOligos(id, df = ref, id_column = "ProteinID",
  oligo_seq = "PROBE_SEQUENCE")
}
\arguments{
\item{id}{a proteinId of interest (character vector of length 1)}

\item{df}{data frame that maps proteinIDs to oligo sequences}

\item{id_column}{\code{df} column with proteinIDs (character vector of length 1)}

\item{oligo_seq}{\code{df} column with oligo sequences (character vector of length 1)}
}
\value{
The function returns a tibble with two columns: proteinIDs and sequence
of concatenated oligos that are associated with the proteinID
}
\description{
Find (and concatenate) oligo sequences that belong to a proteinID of interest.
}
\note{
The function accepts as an input only one proteinID. For more than one
proteinIDs, combine with purrr::map_dfr
}
\examples{
cOligos(49766)
cOligos("43120")

# more than one proteinIDs
purrr::map_dfr(c("49766", "59238"), cOligos)

}
