% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref}
\alias{ref}
\title{A Dataframe with Oligo (Probe) Sequences and ProteinIDs to Which the Oligos Map}
\format{A data frame with columns:
\describe{
\item{ProteinID}{\emph{Spathaspora passalidarum} NRRL Y-27907 protein ID number for
version 1 of the genome assembly, as available on the genome portal
\url{http://www.jgi.doe.gov/spathaspora/}}
\item{PROBE_ID}{ID of the oligo (probe) that is present on the microarray}
\item{PROBE_SEQUENCE}{Sequence of the oligo (probe)}
}}
\source{
Gene Expression Omnibus.
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GPL11079}
}
\usage{
ref
}
\description{
A Dataframe with Oligo (Probe) Sequences and ProteinIDs to Which the Oligos Map
}
\examples{
\dontrun{
ref
}
}
\keyword{datasets}
