% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sp}
\alias{Sp}
\title{ExpressionSet: \emph{Spathaspora passalidarum} Grown on Glucose and Xylose}
\format{An ExpressionSet: Expression profiling by genome tiling array.

Investigation of whole genome gene expression level changes in
\emph{Spathaspora passalidarum} NRRL Y-27907 grown aerobically in xylose, compared
to the same strain grown aerobically in glucose.

A six array study using total RNA recovered from three separate
cultures of Spathaspora passalidarum NRRL Y-27907 grown in glucose and three
separate cultures of Spathaspora passalidarum NRRL Y-27907 grown in xylose.
Each array measures the expression level of 362,487 probes (average probe
length 54.5 +/- 4.0 nt) tiled across the Spathaspora passalidarum NRRL Y-27907
genome with a median spacing distance of 29 nt. During data processing,
probes are filtered to include only those probes corresponding to annotated
protein-coding genes.}
\source{
Gene Expression Omnibus.
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE24853}

Wohlbach DJ, Kuo A, Sato TK, Potts KM et al. Comparative genomics
of xylose-fermenting fungi for enhanced biofuel production. Proc Natl Acad
Sci U S A 2011 Aug 9;108(32):13212-7. PMID: 21788494
\url{https://doi.org/10.1073/pnas.1103039108}
}
\usage{
Sp
}
\description{
ExpressionSet: \emph{Spathaspora passalidarum} Grown on Glucose and Xylose
}
\examples{
\dontrun{
Sp
}
}
\keyword{datasets}
